requirements.txt
asgiref==3.8.1
Django==5.1.4
django-ckeditor==6.7.0
django-js-asset==3.0.1
django-tinymce==4.1.0
pillow==11.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
distro==1.9.0
Django==5.1.4
django-ckeditor==6.7.0
django-js-asset==3.0.1
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
exceptiongroup==1.2.2
google-api-core==1.34.1
google-auth==2.38.0
google-cloud-dialogflow==2.39.0
googleapis-common-protos==1.68.0
groq==0.18.0
grpcio==1.70.0
grpcio-status==1.49.0rc1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
joblib==1.4.2
numpy==2.2.3
pandas==2.2.3
pillow==11.1.0
proto-plus==1.26.0
protobuf==3.20.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
reportlab==4.3.1
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
